+++ src/main/java/controllers/member/LoginController.java
@@ -1,20 +1,17 @@
-package controllers.works;
+package controllers.member;
 
 import commons.ViewUtils;
 import controllers.Controller;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 
-public class SaveController implements Controller {
-    @Override
+public class LoginController implements Controller  {
     public void get(HttpServletRequest req, HttpServletResponse resp) {
-        String URI = req.getRequestURI();
-        String mode = URI.indexOf("edit") != -1 ? "edit" : "add";
 
-        ViewUtils.load(req, resp, "works", mode);
+
+        ViewUtils.load(req, resp, "member", "login");
     }
 
-    @Override
     public void post(HttpServletRequest req, HttpServletResponse resp) {
 
     }
