+++ src/main/java/controllers/works/WorksController.java
@@ -29,7 +29,7 @@ public class WorksController extends HttpServlet {
 
     @Override
     public void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        req.setAttribute("StatusList", Status.getList());
+        req.setAttribute("statusList", Status.getList());
         router.route(req, resp, "works");
     }
