+++ src/main/java/testdao/Ex01.java
@@ -1,9 +1,5 @@
 package testdao;
 
-import models.works.Status;
-import models.works.Work;
-import models.works.WorkDao;
-
 public class Ex01 {
     public static void main(String[] args) {
         /**
