+++ src/main/java/testdao/Ex02.java
@@ -4,7 +4,6 @@ import controllers.works.WorkServiceManager;
 import models.works.SaveService;
 import models.works.Status;
 import models.works.Work;
-import models.works.WorkDao;
 
 public class Ex02 {
     public static void main(String[] args) {
