+++ src/main/resources/META-INF/maven/archetype.xml
