+++ src/main/resources/archetype-resources/pom.xml
