+++ src/main/resources/archetype-resources/src/main/java/App.java
