+++ src/main/resources/archetype-resources/src/test/java/AppTest.java
