--- src/webapp/WEB-INF/view/works/_form.jsp
@@ -1,19 +1,0 @@
-<%@ page contentType="text/html; charset=UTF-8" %>
-<dl>
-    <dt>작업 상태</dt>
-    <dd>
-
-    </dd>
-</dl>
-<dl>
-    <dt>작업 제목</dt>
-    <dd>
-        <input type="text" name="subject">
-    </dd>
-</dl>
-<dl>
-    <dt>작업 내용</dt>
-    <dd>
-        <textarea name="content" id= "content"> </textarea>
-    </dd>
-</dl>
